requirements.txt
asgiref==3.6.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
dj-database-url==2.0.0
Django==4.2.1
django-db-storage==1.0.2
dlib==19.24.1
face-recognition==1.3.0
face-recognition-models==0.3.0
fonttools==4.39.4
importlib-resources==5.12.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
opencv-contrib-python-headless
opencv-contrib-python
opencv-python
packaging==23.1
Pillow
pyfingerprint==1.5
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
python-decouple==3.8
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
zipp==3.15.0
gunicorn==20.1.0
psycopg2-binary==2.9.6
boto3
django-storages
django-jazzmin
django-heroku
redis
celery
django-celery-beat
channels
djangorestframework